refactor(core): extract registerPointerListeners
Pull the 8 pointer/click cluster listeners (mark-copy-waiting copy,
keypress enter-block, pointermove/down/up, contextmenu, pointercancel,
click, copy-preventDefault) out of init(). They coordinate the drag
handlers, prompt-mode commit, toolbar menu dismissal, and context-menu
opening.

Removed 11 now-unused imports/destructures from index.tsx.

index.tsx: 974 -> 801 (-173). 64 pointer + drag + context + touch
specs green.
